+++ test/puppeteer/functions.js
@@ -41,6 +41,23 @@ exports.check_invoice_amounts = async function( page, order_id, excl_tax, tax, i
     expect( data.verollinen ).toBe( incl_tax );
 }
 
+exports.check_invoice_row_amounts = async function( page, order_id, correct_rows ) {
+    await page.goto( config.wordpress_url+"/wp-admin/admin.php?page=wc-orders&action=edit&id="+order_id+"&laskuhari_action=get_invoice_data" );
+    await page.waitForSelector( "pre" );
+
+    const data = await page.evaluate(() => {
+        return JSON.parse( document.querySelector( "pre" ).textContent );
+    });
+
+    const rows = data.laskurivit;
+
+    for( i in correct_rows ) {
+        for( const key in correct_rows[i] ) {
+            expect( Math.round( rows[i][key] * 100 ) / 100 ).toBe( correct_rows[i][key] );
+        }
+    }
+}
+
 exports.logout = async function( page ) {
     await exports.wait_for_loading( page );
     await page.hover( "#wp-admin-bar-my-account" );
@@ -91,7 +108,7 @@ exports.fill_out_checkout_form = async function( page, testid ) {
     await page.click( "#billing_address_1" );
     await page.keyboard.type( "Testroad 123" );
     await page.click( "#billing_postcode" );
-    await page.keyboard.type( "123456" );
+    await page.keyboard.type( "00100" );
     await page.click( "#billing_city" );
     await page.keyboard.type( "Testplace" );
     await page.click( "#billing_phone" );