+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.12.2] 2024-08-15
+
+### Fixed
+
+- Improved error handling for Laskuhari API requests
+
+### Changed
+
+- Updated Finvoice operators list
+
 ## [1.12.1] 2024-05-20
 
 ### Fixed